requirements_hf.txt
streamlit>=1.29.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
pypdf>=6.0.0
chromadb>=1.0.0
numpy>=1.24.0
requests>=2.31.0
huggingface-hub>=0.16.0